+++ webknossos/tests/dataset/test_dataset_download_upload_remote.py
@@ -133,7 +133,7 @@ def test_remote_dataset(tmp_upath: UPath) -> None:
     )
 
     assert not remote_ds.read_only
-    assert not remote_ds.get_color_layers()[0].read_only
+    assert remote_ds.get_color_layers()[0].read_only
     assert remote_ds.get_color_layers()[0].get_finest_mag().read_only
 
     assert remote_ds.name == "test_remote_metadata"