+++ webknossos/tests/test_cli.py
@@ -465,9 +465,7 @@ def test_export_tiff_stack(tmp_upath: UPath) -> None:
             (bbox.topleft.x, bbox.topleft.y, bbox.topleft.z + data_slice_index),
             (bbox.size.x, bbox.size.y, 1),
         )
-        tiff_path = destination_path / _make_tiff_name(
-            "test_export", data_slice_index + 1
-        )
+        tiff_path = destination_path / _make_tiff_name("test_export", data_slice_index)
 
         assert tiff_path.is_file(), f"Expected a tiff to be written at: {tiff_path}."
 
@@ -521,7 +519,7 @@ def test_export_tiff_stack_tile_size(tmp_upath: UPath) -> None:
             for x_tile_index in range(ceil(bbox.size.x / tile_bbox.size.x)):
                 tiff_path = (
                     destination_path
-                    / f"{data_slice_index + 1}"
+                    / f"{data_slice_index}"
                     / f"{y_tile_index + 1}"
                     / f"{x_tile_index + 1}.tiff"
                 )
@@ -586,7 +584,7 @@ def test_export_tiff_stack_tiles_per_dimension(tmp_upath: UPath) -> None:
             for x_tile_index in range(ceil(tile_bbox.size.x / tile_bbox.size.x)):
                 tiff_path = (
                     destination_path
-                    / f"{data_slice_index + 1}"
+                    / f"{data_slice_index}"
                     / f"{y_tile_index + 1}"
                     / f"{x_tile_index + 1}.tiff"
                 )